requirements.txt
python-jobspy>=1.1.70
gspread>=6.0.0
google-auth>=2.27.0
pyyaml>=6.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
pandas>=2.0.0
